backend/requirements.txt
asgiref==3.9.1
autopep8==1.6.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cgi-tools==0.0.4
charset-normalizer==3.4.2
config==0.5.1
decouple==0.0.7
dj-database-url==3.0.1
Django==5.2.5
django-cgi==0.1.2
django-cors-headers==4.4.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.0.0
drf-yasg==1.21.11
git-filter-repo==2.47.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql==0.0.3
mysqlclient==2.2.7
packaging==25.0
pillow==10.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
PyMuPDF==1.26.0
python-dotenv==1.1.1
pytz==2021.3
PyYAML==6.0.3
reportlab==4.4.3
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
termcolor==3.1.0
toml==0.10.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
